Guarding against potential leaks in calls to endTransaction() for prepared queries.
In the generated code logic of a prepared query, a finally block is generated with a call to endTransaction() before a query is released. In the case where endTransaction() call fails, the query will not be released resulting in a potential leak. This change conditionally adds a `try-finally` block wrapping the call to `beginTransaction()` and `endTransaction()` in the case where the query will be released. The `finally` block will release the query regardless of whether or not the transaction fails in the `try` block.

Bug: 254343477
Test: DaoWriterTest.kt
Change-Id: I77aa381324045f5f4d34e8c3edb3aef0b5d5d96b