[M3] Move SwipeableV2 to use density and DismissState and SheetState to use px based thresholds
We are moving the density dependency to the component level.
This applies to the following components: SwipeToDismiss and Sheet based components.

Relnote: We are moving the density dependency to the component level. This applies to the following components: SwipeToDismiss and Sheet based components. Please use the new overload provided where density is a parameter.

Test: Existing tests
Fixes: 277049185
Fixes: 277048426
Fixes: 277050264
Fixes: 277050255

Change-Id: I1846ea6aeb659f53eb8bff6895f7aea19af46fe8