Propagate server job cancellation back to clients.
Currently the CancellationException thrown by SDKs is transformed into a
PrivacySandboxException, which doesn't inherit from
CancellationException so it doesn't properly cancel the app coroutine.
In this change we fix this by introducing a
PrivacySandboxCancellationException that is thrown when SDK exception is
a CancellationException. This keeps the cancellations in sync without
major changes to our binder protocol.

Test: ./gradlew :privacysandbox:tools:tools-{apigenerator,apicompiler,core}:test
Change-Id: Ife460906ec11317a4bcf27cc1f33213f6ccfb559