Do not call requestLayout() during lazy list scrolling when the size didn't change
In aosp/2384214 we started calling requestLayout() when the node remeasure request could potentially change the size of the view to not lag one frame. However it is not needed to call requestLayout() when we do a synchronous Remeasurement.forceRemeasure() as the remeasuring itself will happen straight away after calling requestRemeasure(). This allows us to only call requestLayout() when the size did really change as a result of the remeasurement.

Fixes: 278527662
Test: new tests in ResizingComposeViewTest
Change-Id: I914fc9a5d205100d707638fb61b08dfc3ac5cc51