Allow DefaultSurfaceProcessor to rotate image for ImageCapture snapshot
In ImageCapture, Camera HAL rotates the image based on the CaptureRequest option. Ideally it should be the same for virtual camera, but we were taking a short path and did not implement that. Instead, we never rotate the buffer and special case the exif in ProcessingInput2Packet.java. This works for StreamSharing, however, it breaks for extensions because extensions still rotates the buffer.

In this CL, we update StreamSharing to match the HAL behavior and rotate the image when requested. This keeps the pipeline consistent.

- Update DefaultSurfaceProcessor#snapchat to accept a rotation degree. Update the transformation with the rotation degrees.
- Plumb the rotation degree from CaptureConfig to DefaultSurfaceProcessor.
- Remove the isVirutalCamera bit in ProcessingNode since it's no longer needed.

Bug: 279714190
Test: manual test and ./gradlew bOS
Change-Id: Ic1182e76510ff48a16aa5565437df188ad229c96