Fix @JvmSuppressWildcards issues in KSTypeVarianceResolver.
Unfortunately, there are a number of ways that KSP can lose
`@JvmSuppressWildcards` and `@JvmWildcard` annotation information
as we're transforming types.

For example, using any of the following will drop the annotation:

  * `KSType#replace(List<KSTypeArgument>)`
  * `KSPropertyDeclaration#asMemberOf(KSType)`
  * `KSFunctionDeclaration#asMemberOf(KSType)`

In addition, creating a new type argument will also drop the annotation:

  * `Resolver#getTypeArgument(KSType, Variance)`

This CL tries to handle these cases.

In addition, this CL handles cases where the wildcard annotation is a
typealias itself, e.g. `typealias JSW = JvmSuppressWildcards`.

Test: KspTypeNamesGoldenTest.kt
Change-Id: I00ba3ab692e8ce4a33cc7ee9cfb168f904d909a6