Wrap all AIDL methods with exception logging.
Remove any other log-and-rethrow except toApiComplicationData which adds
the failed ComplicationData to the log (although that should be
converted to a wrapping exception instead).

Most log-and-rethrow were only referenced by AIDL methods, but not all.
Some errors will just propagate to the caller who would likely log the
error anyway.

Bug: 273520594
Change-Id: I0ae46d84b9ac08fd7e6cb9d0fab63957c6dca483