Fix ConcurrentModificationException exception when opening the camera with VideoCapture
Root cause: SessionConfig and CaptureConfig use the same List
instance as in their builder. If builder is later invoked to
add or remove some callback, it could trigger this
ConcurrentModificationException if change happens during
iterating the list.

Solution: Copy the list when building the instance.

Relnote: "Fixed an issue where ConcurrentModificationException
is thrown during camera opening when VideoCapture is bound."
Bug: 274051284
Test: CaptureConfigTest / SessionConfigTest

Change-Id: Ic8ac4a339a45f9f09d0ae0cb20fd4bef19d27cbf