Remove unnecessary calls from magnifier center calculation
While calculating the center of magnifier in Android, we tried to conver the offset between inner text field coordinates and decoration box coordinates. However, as evident by the oversight in code, this was unnecessary. The layoutCoordinates objects used in this code were always referentially the same. So the calls to translate offset between layoutCoordinates objects were doing nothing anyway.

Test: Existing tests
Change-Id: I54d3d19cc1c79140861bca930617ed2eb7ab8b4c