Add additional diagnostic information when snapshots are not open.
The exception generated when a snapshot not open when it is applied does
not provide enough information to determine if it is the result of a
user error or some internal issue in the snapshot system. This change
provides additional information. If the snapshot has been applied
already or disposed then the `apply()` was called twice and this error
is expected. However, if the snapshot has not been applied or has not
be disposed then this might be indicitive of an underlying bug in the
snapshot system.

BUG: 282790642
Test: ./gradlew :compose:r:r:tDUT
Change-Id: I23e1d85d9a9274d76fe1bd2a04ec1d0862b838a7