Add timestamps to predicted events
Following the behavior of the original prediction library, the
predicted events did not contain the down event or event time.
However, this information can be useful when computing the strokes.

This CL also removes a path that would add predictions if it was
not possible to provide as many predictions as requested.

Bug: 283093719
Test: added logging to the sample app to monitor the data, and
the timestamps are correctly populated
Relnote: "Predicted motion events now report the correct down and event time"

Change-Id: I400595643e50e8cdd5c2686df65ce4cd7965b598