Update PublishLayoutInfoTracker to support UiContext.
There was a bug where if you used the Context based methods instead of
the Activity based methods for the PublishRule the value would not be
overriden. This updates the publish rule so that if you override a value
then the override also applies to Context based APIs.

Relnote: """Update WindowLayoutInfoPublishRule to support overrides on
Context based WindowLayoutInfo."""

Bug: 284353680
Test: ./gradlew window:window-testing:cAT
Change-Id: I2037aeb49b5ec6193457bbc11cd04ff56c8b764d