Update BringIntoViewCalculator API to allow a custom animation.
Even though we control how the animation happens, it would be interesting for certain uses cases to use a different animation spec that is retargetable (AnimationSpec). Update the name of the API to BringIntoViewScroller since this is no longer a calculator only.

Relnote: Introduced scrollAnimationSpec to allow custom animation specs. Updated BringIntoViewCalculator to BringIntoViewScroller.

Test: Previous tests should pass
Change-Id: Idb741a8f55df781cc93448159dace9f339aec7b4