[GH] Disable haptics on github
We are hitting b/285039694 on github. Even though we can workaround it, the workaround is adding a dependency on core-ktx from the samples project, which is not a great solution (it needs prebuilt, not module).

Bug: 285039694
Test: Github CI

This is an imported pull request from https://github.com/androidx/androidx/pull/552.

Resolves #552
Github-Pr-Head-Sha: 4b7f9d44729d7f7e0f31da70f96e67c44931cddf
GitOrigin-RevId: 4a724fc5002d9e41b9a50d9c604907a09aa4c988
Change-Id: Ie9762beb99d37427b83c777b08d3dcae533910cc