Update PlatformTimeUpdateNotifier per API feedback
Change parameter type in setReceiver from Callable to Supplier.

This is because calling `Callable#call()` by signature can throw checked
exception, even if we don't do it in the implementation and on the
caller side. By switching to the Supplier, we don't  need try-catch
block in the default implementation.

Also includes:
- Clarify Javadocs for setReceiver
- Log exception from future in default impl

Bug: 275682976
Test: N/A
Relnote: "We have changed the parameter type in the `PlatformTimeUpdateNotifier#setReceiver` from Callable to Supplier."
Change-Id: I664bf13fb47a8da7b8e0348d5d8cee5a2e14dbe4