Fixed occasional flicker with Canvas front -> multi buffered layer handoff.
Updated MultiBufferedCanvasRenderer to
maintain a collection of outstanding
Image/HardwareBuffer instances and
include a release method for consumers to
invoke to return resources back to the
underlying ImageReader. Updated Image
usages to leverage [set|get]Fence on
appropriate API levels.

Refactored CanvasFrontBufferedRenderer to
have the front/multi buffered SurfaceControls
be siblings instead of the front buffered layer
be a child of the multi buffered layer.
Additionally only apply pre-rotation to the
front buffered layer as the flicker is seen
when updating the multi buffered layer with
the configured buffer transform.

Updated MultiBufferedCanvasRenderer usages to
specify usage flags and configure a preservation
flag to have the underlying HardwareRenderer
instance conditionally clear content based on
visibility changes of the front buffered layer

Refactored CanvasFrontBufferedRenderer to remove
all usages of onCommitListeners

Fixes: 285005324
Fixes: 282794011
Test: Added test to MultiBufferedCanvasRendererTest
Change-Id: I6c7562212119e6ca47092b005f0a37c08e5e3430