Refactor BTF2 internals to consist of 3 Modifier Nodes.
BasicTextField2 in high level design is only 2 parts; decorator and inner text field.

First, decorator is mostly responsible for styling, placement of auxiliary composables. In other words it's a reverted slot API. Decorator also plays as the overall gesture detection area to gain focus, initiate input session with IME, move the cursor and selection, enable scrolling of its internals.

Second, core is the actual text editor which needs to be placed somewhere in decorator. Core's main functionality is to layout, place, and draw the text that the user actually sees.

However, Core can be split into two parts. In the current definition, core has the role of both viewport and the actual content that's drawn behind the viewport. This may cause confusion when we want to separate between the scrolled, cropped, clipped Node that is drawn on the screen and the rest of the text layout Node that can still be scrolled into view.

To make it easier to map gesture offsets from decoration box onto actual text layout result (for querying text indices where touch happens etc.) it'd be better to create a new TextLayoutModifierNode that reports its LayoutCoordinates separately from the rest of the core modifiers.

Test: mostly existing tests
Change-Id: If6ea73861fe64d09c46a1b74ba941f8c6f5ad2e7