Improve SteppedForLoopDetector
1. Narrow down the detection

Stepped loops with `..` are properly optimized by kotlinc.
Instead it's loops using `until` or `..<` that are not.
See https://youtrack.jetbrains.com/issue/KT-59115

2. Remove the usage of deprecated APIs

PsiType.INT and PsiType.LONG are now deprecated, instead let's
use the canonical text representation of PsiType to check when
the type is an int or a long.

In newer versions of PSI we will be able to use
PsiTypes.intType() and PsiTypes.longType() instead.

Bug: 285702293
Bug: 285968541
RelNote: N/A
Test: SteppedForLoopDetectorTest
Change-Id: I1379ab6548e62c7889149cb4efed07bd794298bf