Add an experimental composition observation API
The change adds and API that can be used to determine why a comppostion
was invalidated. This change introudces two new experimental observers,
the CompositionObserver, which observes the beginning and ending of all
compositions and recompositions of a Composition. And
RecomposeScopeObserver which will observer the beginning and ending of
recomposing a recompose scope which loosly corresponds to a composable
function starting and ending. These, combined with the current tracing
API, can be used to correlate changes to state object to the
corresponding scopes that where required to be re-evaluated.

Test: ./gradlew :compose:r:r:tDUT
Fixes: 286287720
Relnote: """Added an experimental composition observation API that
allows observing when compositions and recompose scopes are exeucted
and provides information about what caused them to be invalidated."""

Change-Id: If5339e1402fd40f14892f0c7746f6b4ad1fe5687