Fix index out of bounds issue with parameterized receiver parameters.
The issue here was that we were passing the wrong parameter index when
creating the MethodParameter scope because we were not taking the
receiver parameter into account. This CL fixes the issue by shifting the
parameter index by one when a receiver parameter exists.

I've also updated the `toString()` behavior for
`KspSytheticContinuationParameterElement` and
`KspSyntheticReceiverParameterElement` to return the parameter name,
which matches the behavior of `KspParameterElement`.

Test: XExecutableElementTest.kt

Change-Id: I6bb76aa2a601aa384ec2d9c06f1e1b7d86f50abe