Adding Wear Compose Material 3 Inline Slider
InlineSlider allows users to make a selection from a range of values.
The range of selections is shown as a bar between the minimum and
maximum values of the range, from which users may select a single value.
InlineSlider is ideal for adjusting setting such as volume and
brightness.
Value can be increased and decreased by clicking on the increase and
decrease buttons, located accordingly to the start and end of the
control. Buttons can have custom icons - decreaseIcon and increaseIcon.
The bar in the middle of control can have separators if segmented flag
is set to true. A single step value is calculated as the difference
between min and max values of valueRange divided by steps + 1 value.

Test:  androidx.wear.compose.material3.Slider tests
Bug: 286388266

Relnote: "Adding Material3 InlineSlider to Wear Compose. It allows
users to make a selection from a range of values. The range of
selections is shown as a bar between the minimum and maximum values of
the range,from which users may select a single value. InlineSlider is
ideal for adjusting settings such as volume or brightness."

Change-Id: I7085f3340fcc76556fadb06ff81c598ae84eacdc