[Extensions] Change ClientVersionBackwardCompatibilityTest to remove the callback methods call check.
After introduce the default methods in callback, OEMs will always
invoke the callback methods regardless of the client
Extensions-Interface version. Therefore the test to verify if
some callback methods are invoked or not doesn't make sense.

The tests are changed to test if preview and image capture can work
well for each client versions.

Bug: 287399418
Test: ClientVersionBackwardCompatibilityTest.
Change-Id: I877b118b86b351994c7000b5d1dff7adafcedf98