Fix PreviewParameterProvider can't use value class
Currently, when a PreviewParameter is using sequences of value
classes, it shows a type mismatch error. The cause comes from
java reflections, it compares the given value class with the
unwrapped value, creating a type mismatch.

This change ueses a bit of Kotlin reflection to get
the unwrapped value from the PreviewParameter sequence, fixing
the type mismatch exception.

Test: ComposeInvokerTest, ComposeViewAdapterTest, PreviewParameterTest
Fix: 266554124
Change-Id: I2a6b4def167a090ff4b16b3f369a4550c843f948