Fix Primitive State quickfixes for library users
This CL fixes an issue for projects that depend on the Compose runtime.
Properties annotated with `@AutoboxingStateValueProperty` would
correctly display a warning referenced, but would not correctly resolve
the quickfix replacement and suggest `<unknown replacement>` instead.
Currently, the quickfix only seems to resolve the replacement correctly
from inside the AndroidX repo itself.

This seems to have something to do with how Kotlin compiles property
annotations and how lint runs when it only has access to a compiled
Kotlin library instead of a source module. Annotations targeting
properties don't compile onto the getter or setter, and instead only
appear on a synthetic method. By migrating the annotation to target
both the getter and the setter, Lint seems to have a much better time
finding the annotation since it appears directly on the getter and
setter.

Bug: b/287279257
Relnote: N/A
Test: AutoboxingStateValuePropertyDetectorTest
Test: Manually test in a project that depends on the runtime module
      produced by ./gradlew createArchive
Change-Id: Iba9536e23d6d94f1056cf7859454e4e2d6ee03cb