Introduce UserInitiatedTaskRequest and related classes.
The UserInitiatedTaskRequest object will be the main request class which
will hold relevant data for a user-initiated job. The UserInitiatedTask
class will be used to perform the task asynchronously.

The Constraints object will hold the network constraints for the new
UserInitiatedTaskRequest. In the future, this object will support other
constraints like battery and storage.

Bug: 280127680
Bug: 280127978
Bug: 280127935
Test: ./gradlew :work:work-datatransfer:assemble
Test: ConstraintsTest
Test: UserInitiatedTaskRequestTest
Change-Id: I669bbc35f8bf587cc4317d7153e6171a9185559c