Support AnimateChanges with the Modifier DSL
To provide the same functionality as when using the ConstraintSet
pattern.

Note that it takes a different code path since the ConstraintSet is
obtained until after composition is done.

Relnote: "You may now use `animateChanges = true` when using
ConstraintLayout with the inline Modifier DSL (Modifier.constrainAs),
whenever a change on the constraints DSL is done ConstraintLayout will
automatically animate to that new state."

Bug: 272319113
Test: ConstraintLayoutTest
Change-Id: I9abf1451ef5123ba2a69b794f85ca07aedcbfb10