Fix SurfaceView screenshot black bar
When taking a PreviewView screenshot with PERFORMANCE mode, on Pixel 6 / OS 14, there is a black bar at the top. The root cause is that we don't wait for the screenshot callback before rendering the result. The screenshot API somehow gets slower on Pixel 6 / OS 14, so it leaves a black bar unrendered.

Fix: use Semaphore to wait for the completion of the screenshot.

Bug: 289236539
Test: manual test and ./gradlew bOS
Change-Id: I3b0b9d874319cb5ab8fb21094a068b99292ddc85