Address API feedback
Ensure that the composition observer API will work when wrapped.

Added a service locator for composition services such as the
observing composition and uses this to locate the internal
observer services instead of relying on the composition being a
particular implementation. This allows using observation when
the composition is wrapped as done in the UI module.

Moved the API to `tooling`

Moved the composition observer API to the `tooling`
package to make it more clear that this is only intended to
be used during application development.

Test: ./gradlew :compose:r:r:i-t:cC
Fixes: 288898192
Relnote: """Added a service locator interface that can be
implemented by implementors of Composition that allows an
implementation of Composition that delegates to another
composition delegate service lookups to the original
Composition. This should not be called directy and is used
to enable creating experimental APIs in the runtime that
can be found from wrapped versions of the composer such
as the UI module does."""

Change-Id: I296b9deae0c30b70d5e37f7b5fa992d2583ef4f4