Provide a new way to generate ExpandableStates
This is needed for more complex use cases, like having a series of
expandables that can be added/removed at runtime.
Added sample of ScaluingLazyColumn where each item can be
expanded/contracted and items can be removed and added.

Bug: 289217657
Test: ./gradlew :wear:compose:compose-foundation:connectedAndroidTest
Relnote: "Provide a new way to generate ExpandableStates, for cases in
which the they need to be created on demand, not necesarily on a
@Composable scope"

Change-Id: Iff9e0dc927754c763ec0c1c0b9760ec5a8f20ef5