1.5.0-beta03 merge. Fix CoreText.kt (#664)
- Apply changes from jb-main between the base commit 246a0a09f5335a92f85d3f14597ecddbdec68f9f and the pre-merge commit fd66a3ee078b92771b9f5b7098a5163b2ad1fd8e:

  - dca000a2 (this commit was a squash, and its message is a mess)

The `pointerInput(Unit)` is restarted every recomposition, because we create `mouseSelectionObserver` every time.

It can cause selection issues. We can accidentaly not have issues (they depend on how `mouseDragGestureDetector` is used inside TextField), but it can return in the future, so better to write these modifiers in a proper way.