Update behavior of prefetcher.
Changed the behavior of the prefetcher not to perform prefetching as the page bounces back when the scrolling operation does not exceed the threshold to move to the next or previous screens. The condition is if the animator scrolling direction is the same as the direction of the pointer gesture, only then prefetching is performed.

Bug: 289088847
Test: Added test to verify behavior.
Change-Id: Ic742d3276d443948b76874ae04671b9b024609ad