Re-enable load after store configuratation cache feature
It was originally disabled due to KMP bug in aosp/2376248
We just upgraded to KGP 1.9.0 and this issue should be fixed.

Suppress a warning about No valid plugin descriptors as that
is a bug in Gradle that will be fixed in Gradle 8.3. In the mean
time it will not be validating whether Gradle plugin projects
are including a plugin descriptor META-INFO/gradle-plugins entry.
Since these rarely change, the risk is very low.

Test: None
Change-Id: I086a17087904afa99608bae599da82a88d997347