New enter/exit transition to scale content to container size
- Introduce a new type of enter/exit transition that scales
the content to fit in the resizing container. This allows
AnimatedContent to swap content for target change in a
container-transform style.
This is a prerequisite for shared element based container
transform.

- Added a (not shared element based) container transform
demo

- Made LookaheadScope stable and create a LookaheadScope
in AnimatedContent

Test: tests included

Relnote: "New type of enter/exit transition that scales
the content based on the size of the animating container
during enter & exit animation.

LookaheadScope Composable fun and interface are now stable.
"

Change-Id: Ifb2ce945db06e291aeb0980872b427bf0a580ede