Stop transforming non-composable fun interfaces in Compose compiler
The original bug (b/178663739) is already fixed upstream, so we can remove this transform, as it causes problem with Java sam conversions in K2.
I left the conversion for composable functions, as SAM conversion does not support restartable lambdas otherwise.

The side effect of this change is that SAM-converted lambdas are now memoized by Compose compiler. This CL also adjusts this behavior by ensuring remember happens outside of conversion. Without this fix, the converted class was delegating to remembered lambda, essentially un-doing the optimization.

This commit also improves the lowering for cases when SAM_CONVERSION was followed by IMPLICIT_CAST (see updates in `ComposerParamSignatureTest` as sample).

Test: Updated compiler tests + runtime tests
Bug: 178663739
Fixes: 288403625
Change-Id: Ieabb4a9df24f77e9b5e7e41adf360c50cede604f