Adds auto-mirroring support to the generated icons
- Updated the vector XML parsing to read the `autoMirrored` attribute.
- Generated a new `automirrored` package that holds auto-mirrored icons
  per icon-theme. For example: Icons.AutoMirrored.Filled.ArrowForward.
- Updated the ImageVectorGenerator to generate and place auto-mirrored
  icons at the new package.
- Have the generator mark any existing icons with an auto-mirrored
  version as deprecated, and suggest a replacement with the
  auto-mirrored one.
- Updated the generated test files to include an additional property
  that have the ImageVectors with the auto-mirror flag.
- Updated the tests to run against two groups of icons separately
  (mirrored and unmirrored).
- Split the API files generation to icons.txt and automirroed_icons.txt.
  Update the icons.txt to hold only icons that are not in the
  automirrored_icons.txt file.

Bug: 279989215
Test: Updated the generated test icons, and the tests themselves
Relnote: "Added support for auto-mirrored icons when rendered in
right-to-left layouts. Icons in the material-icons-core and
material-icons-extended modules are now providing additional icon sets
for supporting auto-mirroring when the icon allows it. The new sets are
prefixed with `Icons.AutoMirrored.Filled...` etc., and hold icons that
will auto-mirror on RTL layouts.
See developers.google.com/fonts/docs/material_icons for the list of
icons that can (and should) be auto-mirrored.
The previously provided icon properties for those icons are now marked
as deprecated, and provides a replacement-block suggestion to help with
the migration.
In case you don't have any special handling for mirroring icons on RTL,
we recommend migrating to the new set of icons. For example,
`Icons.Filled.ArrowBack` should be refactored to
`Icons.AutoMirrored.Filled.ArrowBack`."

Change-Id: I4b5116207cb9c0dd1a8a72d2bd3f606f8fdd0129