cfw: Update ComposeWindow implement resizing strategy (#692)
Updated the resize function in ComposeWindow.js.kt to avoid multiple event listeners being applied on every resize event. Implemented via cloning the old canvas and replacing it with the new one. This is a workaround solution and a better strategy to fix this in skiko will be considered in the future.