Pass the suggested transformation to the effects API
Even the effect does not rotate or mirror the input, we need to pass that info to the caller. The caller should know that if the input would be rotated and mirrored later, so that they can understand e.g. which direction is up in the input.

Right now, we are not passing that information to the caller. This CL fixes the issue by creating the input edge with those information. The info was later overriden when configuring the node, so the node's transformation won't change. The only change is that the caller's SurfaceRequest will now get a correct rotation degrees and mirroring for front camrea.

- Set StreamSharing's mirror mode to be MIRROR_MODE_ON_FRONT_ONLY
- When creating the camera edge, pass the intended rotation degrees and mirroring bit.

This is considered a beta bug fix.

Bug: 292254688
Test: manual test and ./gradlew bOS
Change-Id: Iae4c319703d708715d3f841014fe067228c9beba