refactor(non linear font scaling): Move android-dependent tests out of common unit tests
Once we switch to expect/actual for FontScalable in the next CL, the
common unit tests will try to use Android classes from :core:core that
don't exist in common. So we need to move those to Android.

GraphicsLayerScopeTest was extra tricky, because the unit test uses a
stub for SparseArray, which accepts put() items but doesn't actually
save them in the map. So the size always stays 0, and that causes the
static assertion in FontScaleConverterFactory to fail.

Bug: 239736383
Test: ./gradlew :compose:ui:ui:testDebugUnitTest
  && ./gradlew :compose:ui:ui-unit:connectedAndroidTest

Change-Id: If1fd27254c19aa2ff06709cdf79360cc565c963e