Do not add effects as part of OnCommit
We should never be adding effects as part of the onCommit of another
Effect. The onCommit should be the termination point for the processing
of Effects.

For animations this means we should instead add the no-op as part of
onStart as we know what direction the Animation is going then.

For transitions we should just complete the transition immediately.

Test: all tests pass
Bug: 287137044
Change-Id: I8ba1996657409f5357f8826d92c0eadb98be4e82