Add a quirk inside CameraPipe to ensure that CameraExtensionCharacteristics have been loaded before the CameraExtensionSession call is executed
An Android Framework issue has been discovered where if an application does not query for CameraExtensionCharacteristics before trying to create a CameraExtensionSession, the session will fail to create.

This is because Camera2 has a global variable recording if advanced extensions are supported or not, and the variable is updated the first time CameraExtensionCharacteristics are queried. If the variable is not set, Camera2 will fall back to basic extensions, even if they are not supported.

Bug: 293473614

Test: ./gradlew\
 :camera:camera-camera2-pipe:testDebugUnitTest\
 :camera:camera-camera2-pipe-testing:testDebugUnitTest\
 :camera:camera-camera2-pipe-integration:testDebugUnitTest
Change-Id: I3d6acd0baeaae81f405933c45bd88f0123c3e1d0