Call onStart after collecting operations
Now that collecting and running effects have been separated, we should
move the starting of operations to after we collect them.

Starting the operation changes the state of the View and we do not
actually want to do that work until we are ready to start an effect.

With this change, we can no longer early return for animators whose
visibility is unchanged since we used the new state of the view to
determine that value.

Also the SpecialEffectControllerTest needs to be updated to override
commitEffects.

Test: updated tests
Bug: 287137044
Change-Id: Ib52f9daec99684ca3c04412ec59cd0f1b44991dc