Correct how movable content nodes are diassembled.
If more than one movable content is removed from composition
simultaneously the calculation of the node index (the offset
the node is in its parent node) could become incorrect. This
was because the node groups were not special-cased as they
unconditionally count as one node in their parent node
regardless of the number of nodes they contain.

Fixes: 287010398
Test: ./gradlew :composer:r:r:tDUT
Change-Id: I5a12f4174c5ec7cf9809dfae9ac61bfa8f454f4b