Add TilePreview annotation and preview support
This change updates the TileServiceViewAdapter to preview tiles based on
a method instead of a class inheriting TileService.

This change depends on the change
Ia2f255a791d24b81bbed94eae9176156a8a685a7 in Android Studio.

Bug: 227474520
Relnote: Module for wear tiles tooling is updated for tile preview support and is set for being published
Test: manual
Change-Id: I63d0f04780f4f28e67d40951057bcdc814f5f6d8