First steps of making room-runtime KMP.
This CL mainly sets up room-runtime as a KMP project, adding necessary directories for splitting code between commonMain, androidMain and nativeMain. It also updates the relevant gradle files.

This is a starter CL in a row of many more upcoming CLs, but we are looking forward to renovating our Room, even if it means knocking down some walls and adding a few new ones!

Bug: 299168035
Test: Existing integration tests for Room.
Change-Id: I23c90c33e1645385bfb1f55f3843749e4cb76a11