Re-order DocumentModel.SpecialField enum members
The enum member ordering affects the code generated by FromGenericDocumentCodeGenerator in the current setup. There can be a discrepancy in the order in which we initialize properties vs the order in which we set them on the document class.

This discrepancy gets resolved in the upcoming refactor s.t. we initialize and set properties on the document class in the same order.

This minor change is only here to avoid a large number of diffs in the golden files in the upcoming CL.

Test: Unit/golden tests
Change-Id: Ifbb3bb58a9b8b85615efc8c0a6a6dc9b1be79209