Make BTF2 onTextLayout take a () -> TextLayoutResult?
See https://docs.google.com/document/d/1x4N0UyMDijAHbl_ogBm2ahF3lKCZjmsYpy5ctKM5HGk/edit

This is the API change part of the change to make the layout result
calculated on-demand in aosp/2681239. Doing this as a separate CL
because it requires updating a bunch of tests, and I don't want to
update those tests in the same CL as the impl change.

Test: existing tests updated
Relnote: n/a
Change-Id: Iacb6cb6562df02e418c610ebad898fa424dab803