Fix issue with changing state as part of composition
We should not change state as part of composition as that can cause
inconsistent behavior when reading state. We should instead only ever
change state in an Effect.

RelNote: "Fixed an issues in Navigation with Compose where when using a
Scaffold it was possible to get an error trying to access a DESTROYED
ViewModel"
Test: Added test in NavHostTest
Bug: 268422136

Change-Id: I1dc115fc2b454d68c1a23c78e72707d6d4c739a5