Bugfix: blackscreen on LEGACY devices when stream sharing is enabled.
To reproduce the bug: 1) enable stream sharing, 2) press home button and 3) restore the Activity.

The cause is that on LEGACY devices, the camera send down SessionConfig error callback to reset the pipeline when the app is backgrounded. StreamSharing resets itself and all the children's, but doesn't connect to the children after reset, which causes the issue.

The fix is triggering onUseCaseReset in the parent's SessionConfig callback.

Bug: 295114036
Test: manual test and ./gradlew bOS
Change-Id: Ia68a5d123d969e7ae662911515d0be9ed373cf11