Handle multiple files in the same directory for MultiProcess datastore
This CL adds a workaround for an Android bug where, if multiple FileObservers
are added to the same file/directory, only the last one works.

We solve this by creating a custom observer that multiplexes our observers.
It still won't work if the application has other FileObservers but we don't
have any control over that.

Bug: 279997241
Bug: 37017033
Test: MultipleFilesInTheSameDirectoryTest, MulticastFileObserverTest
Change-Id: I5694e82e3f17719526d670f3a7af9227ba19e573