Adjust ViewConfiguration Compat Logic for Fling Velocity
For API versions that do not support the per-device fling velocity
thresholds, we introduced a compat logic. We are now aligning the compat
logic to be similar to the platform logic. That is:
- If specifying threshold for a given device-type is supported, but no
  threshold has been specified, block fling.
- If specifying threshold for a given device-type is supported and there
  is a specified threshold, use that, as long as it's valid. If it's
  invalid, block fling.
- If specifying threshold for a given device-type is NOT supported, use
  the default fling thresholds.

This is important to avoid inadvertently allowing fling for
supported-device types on older Android versions. That is, if these
device-types get NO fling in Android U+ by default, they should also get
NO fling before Android U by default.

Bug: 294895104
Test: ViewConfigurationCompatTest
Change-Id: I4cad22dfacdbb5023e0bc014354af1c4dcdfbbaa